+++ project/main.gd
@@ -4,6 +4,8 @@ extends Node
 
 
 func _ready() -> void:
+	SentrySDK.logger.info("Starting UI on %s", [OS.get_name()])
+
 	if await cli_commands.check_and_execute_cli():
 		# Quit if a CLI command was executed
 		get_tree().quit(cli_commands.exit_code)